productos y carrito

+++ api/productos.js
@@ -6,9 +6,17 @@ const multer = require("multer");
 let router = new Router();
 let archivo = new Contenedor("text.json");
 
-router.get("/productos/", async (req, res) => {
+const elapsed = Date.now();
+const hoy = new Date(elapsed)
+const diaHoy= hoy.toLocaleDateString()
+
+
+
+/******************/
+//a.get
+router.get("/", async (req, res) => {
   let data = await archivo.getAll();
-   
+  
   res.render("index", { data: data });
 });
 
@@ -23,46 +31,65 @@ let storage = multer.diskStorage({
 
 let upload = multer({ storage });
 
-router.get("/form",(req,res)=>{
-  res.render("form")
-})
-
-
-router.post("/form", upload.single("myfile"), (req, res) => {
-
-  let newProduct = {
-    name: req.body.name,
-    price: req.body.price,
-    urlFoto: req.body.urlFoto,
-  };
- 
-  let idProductoAgregado = archivo.save(newProduct);
- // res.send(
- //   `El archivo se guardo correctamente y el id del nuevo productos es ${idProductoAgregado}`
- // );
-  res.render("form")
+router.get("/form", (req, res) => {
+  res.render("form");
 });
 
-router.get("/:id", async (req, res) => {
+/******************/
+//b.post
+router.post("/form", upload.single("myfile"), (req, res) => {
+  if (req.query.admin) {
+    let newProduct = {
+      timestamp : diaHoy,
+      name: req.body.name,
+      description : req.body.description,
+      codigo :req.body.codigo,
+      urlFoto: req.body.urlFoto,
+      price: req.body.price,
+      stock: req.body.stock,
+    };
+    let idProductoAgregado = archivo.save(newProduct);
+    alert(`se guardo el producto ${idProductoAgregado}`)
+    res.render("form");
+  } else {
+    res.send("error no esta autorizado apra acceder");
+  }
+});
+/******************/
+//get:/':id?
+
+router.get("/productos/:id", async (req, res) => {
   let data = await archivo.getById(req.params.id);
   res.json(data);
 });
 
-router.delete("/:id", async (req, res) => {
-  let data = await archivo.deleteById(req.params.id);
-  res.json(data);
-  //res.send(productos);
+/******************/
+//c.put
+
+router.put("productos/:id", async (req, res) => {
+  if (req.query.admin) {
+    let newProduct = {
+      name: req.body.name,
+      price: req.body.price,
+      img: req.body.img,
+    };
+
+    let data = await archivo.putById(req.params.id, newProduct);
+    res.json(data);
+  } else {
+    res.send("error no esta autorizado apra acceder");
+  }
 });
 
-router.put("/:id", async (req, res) => {
-  let newProduct = {
-    name: req.body.name,
-    price: req.body.price,
-    img: req.body.img,
-  };
-
-  let data = await archivo.putById(req.params.id, newProduct);
-  res.json(data);
+/******************/
+//c.delete
+router.delete("/productos/:id", async (req, res) => {
+  if (req.query.admin) {
+    let data = await archivo.deleteById(req.params.id);
+    res.json(data);
+  } else {
+    res.send("error no esta autorizado apra acceder");
+  }
 });
 
 module.exports = router;
